Changes to load Dictionary attributes from buffer
Code Changes fix this reported issue - https://github.com/radcli/radcli/issues/24

Synopsis:
Currently, the dictionary configs have to be provided via a file and
location of file is not dynamic (statically programmed). This makes the
code not much portable and flexible.

With this change, the Dictionary contents can be provided via a buffer.
Sample use-case is provided in radembedded_dict.c.