Support for non-temporary IPv6 address
Description:
In Linux/BSD/Apple-OS, it is possible to enable temporary IPv6 addresses in
system by default (when IPv6 Privacy Extensions are enabled in system). In this
scenario, with default settings (if bindaddr option is set to "*" in config
file), the radius client will use the temporary ipv6 address as source address
for the packets sent out. This is usually not desired since the RADIUS Server
identifies the RADIUS Client usually by a unique address and continuously
changing IPV6 addresses requires changes in RADIUS Server config too which is
not preferable.

Fix:
In these cases, it is preferable to make Radius Client to use Global IPv6
address as source address for packets sent out.

Changes made:
Introduced a config file option to specify support of IPV6 non-temporary
address support. If this option is set to "true" in config file, Radius Client
will use Global IPv6 address as source address for packets sent out.