Added support for IPv4 and IPv6 source-interface Source-interface on Radius :
When Radius login authentication happens the RADIUS IP packets from the device goes, the source IP  of the packet will have the IP of the egress interface. There could be many egress interfaces on the device. The IP packets can go out on any of the egress interfaces. In that case, the RADIUS IP packet could end up having different source IP’s.  This would cause difficulty while parsing the logs when the same radius client sends the packet with different source IP’s.

To avoid this, we can configure “source-interface” for the RADIUS. This would allow the RADIUS IP packet to have the same source-ip (i.e the IP of the source-interface configured on the RADIUS).
Current support:

Current “radcli” implementation uses “bindaddr” from the radcli.conf file to establish a socket conection to send the RADIUS packets. If the “bindaddr” is not set, any IPv4 egress interface is used.

Enhancement:
Extend it to support both IPv4 and IPv6. When there are mix of IPv4 and IPv6 radius servers configured, the radcli needs to choose the source-ip as either IPv4 or IPv6 respectively based on the radius-server IP version.
Introduced another attribute in the radcli.conf “bindaddr_v6” which will hold the IPv6 source-ip address if configured. When the radcli loops through radius-servers to authenticate,
based on the IP version of the radius-server, the bindaddr/bindaddr_v6 would be selected.